fix inferReciprocalComponent called on unsaved
DataObject::inferReciprocalComponent  cannot be called on object without IDs. There is currently no checks if owner has a proper id, which could result on errors.
Also use $owner instead of $this->owner for consistency in the function.

to be honest, I'm not sure how I managed to do that, but this extra check cannot hurt in my opinion.