FIX Add versioned "latest version" subqueries as a condition in some cases
Currently the "latest version" subquery is applied as an inner join. This is an effective way to filter the list of records being pulled as one query is done for all records returned in the query. However, this doesn't scale well as the _Versions table increases in size. In this case it's much more performant to apply the subquery as a condition. This is run once for each record being returned but is filtered by the primary key so turns out to be quicker than the inner join