FIX Revert framework constraint bump and add easly returns when non B/C interface does not exist
This change must be released as a 1.3.x patch tag, then the Composer constraint can be bumped
back again to where it was. This will fix cases where Composer installs silverstripe/versioned
~1.3.0 on older SilverStripe 4.0/4.1 projects, which has a backwards compatibility break in it.

The change here allows the code to be backwards compatible with older versions. Once tagged
it can be bumped and re-tagged in future in the same release line without concerns for the
same B/C break.