FIX archive relationships, not related objects
The archive action is set to take the place of the delete button,
considering that when a published item is deleted (i.e. from the
draft table) then it cannot be unpublished as no interface exists
to do so.

However this was having the unintended affect of removing objects
people wished to only remove from a relationship list, as the
archive action did not account for whether the delete action intends
to remove the related object, or the link to it (the relationship).

This is most easily remedied by providing the user with both options.
It may not be the best user experience, but is the most flexible.