NEW migration task to publish unVersioned objects
DataObjects with the Versioned extension may not have always had this
setting applied to them. What was previously "always live" by nature of
not having a "draft" version now are by default in draft, as it is the
"base table" that is used as the draft with the Versioned extension.

We need to migrate all objects that exist in the draft table and have no
recorded version at all to the live table, in order that they may continue
to function as they had before the Versioned extension was applied.

This is a new build task that does exactly that. It searches for any
DataObject in this state, and publishes it - but care is taken to not
cascade that publish to cause issues with any new owns relationship
that may have been set up.