FIX: Decouple test setup.
Decoupling modules so that versioned doesn’t rely on an arbitrary and
undefined subset of the CMS recipe nudges our architecture in the right
direction. A more successful decomposition of the recipe into packages
would ensure that there aren’t bidirectional or cyclic relationships
between packages, and instead they are arranged more into “layers”.

This patch means that you can test the module independently, like so:

 * (clone module)
 * composer install
 * vendor/bin/phpunit

The test suit not longer relies on HTMLEditorField, which in practise
is not fully functional without admin. It also skips 

There is 1 test that appears to test a piece of LeftAndMain behaviour.
Arguably it should be refactored and placed there, but I’ve made it
skippable instead.

With that in place it tightens the require-dev dependencies to only
include graphql, as some of the test suite tests the graphql/versioned
behaviour.

I haven’t changed the travis.yml configuration as is appears these are
being auto-generated to be consistent across all modules that make up
the CMS recipe (44fb86acffcd3f4041b03f12299b07348439b2f0).